+++ app/helpers/project_information_helper.rb
@@ -1,0 +1,7 @@
+module ProjectInformationHelper
+  def age_range(establishment)
+    return nil if establishment.age_range_lower.blank? || establishment.age_range_upper.blank?
+
+    "#{establishment.age_range_lower} to #{establishment.age_range_upper}"
+  end
+end