+++ app/models/academies_api/establishment.rb
@@ -1,10 +1,21 @@
 class AcademiesApi::Establishment < AcademiesApi::BaseApiModel
-  attr_accessor :local_authority, :name
+  attr_accessor(
+    :name,
+    :local_authority,
+    :type,
+    :age_range_lower,
+    :age_range_upper,
+    :phase
+  )
 
   def self.attribute_map
     {
+      name: "establishmentName",
       local_authority: "localAuthorityName",
-      name: "establishmentName"
+      type: "establishmentType.name",
+      age_range_lower: "statutoryLowAge",
+      age_range_upper: "statutoryHighAge",
+      phase: "phaseOfEducation.name"
     }
   end
 end