+++ spec/helpers/project_information_helper_spec.rb
@@ -1,0 +1,29 @@
+require "rails_helper"
+
+RSpec.describe ProjectInformationHelper, type: :helper do
+  describe "#age_range" do
+    context "when the lower age range is nil" do
+      let(:establishment) { build(:academies_api_establishment, age_range_lower: nil) }
+
+      it "returns nil" do
+        expect(helper.age_range(establishment)).to be_nil
+      end
+    end
+
+    context "when the upper age range is nil" do
+      let(:establishment) { build(:academies_api_establishment, age_range_upper: nil) }
+
+      it "returns nil" do
+        expect(helper.age_range(establishment)).to be_nil
+      end
+    end
+
+    context "when both upper and lower ranges have values" do
+      let(:establishment) { build(:academies_api_establishment) }
+
+      it "returns a formatted date range" do
+        expect(helper.age_range(establishment)).to eq "11 to 18"
+      end
+    end
+  end
+end